Added missing methods for visibility handling of Home button area.
There was a method to make mHomeLayout invisible(/"gone") but nothing to
determine its current visibility or toggle it. This is useful when
wanting to use the same ActionBar across several activities but switch
the Home button/logo on/off depending on context (in the normal sense
and the Android API sense), for example.